rename: 명확하게 컴포넌트 이름 변경
