+++ pages/[userEmail]/[courseId]/todo/[todoId]/write/index.tsx
@@ -2,7 +2,7 @@ import { NextPage } from "next";
 import { useRouter } from "next/router";
 import React, { useEffect, useState } from "react";
 import { useRecoilValue } from "recoil";
-import UserCheck from "../../../../../../components/UserCheck";
+import UserCheck from "../../../../../../components/QueryStringUser";
 
 import CourseHeader from "../../../../../../domain/Course/Detail/CourseHeader";
 import EditerMaster from "../../../../../../domain/Editer/EditerMaster";
