+++ pages/[userEmail]/index.tsx
@@ -11,7 +11,7 @@ import TodayTodoList from "../../domain/Todo/TodayTodoList";
 import CourseList from "../../domain/Course/CourseList";
 
 import useGetString from "../../hooks/useGetString";
-import UserCheck from "../../components/UserCheck";
+import UserCheck from "../../components/QueryStringUser";
 
 const getTabComponent = (key: string) => {
   switch (key) {
